fix(google): Move google login function to utils folder and out of Landing and Signup Screen
