+++ packages/app/utils/constants.js
@@ -1,3 +1,3 @@
-const URL = "https://4ac9-2600-1700-7860-56a0-7d31-a507-62c3-9f58.ngrok-free.app";
+const URL = "https://04aa-107-198-190-216.ngrok-free.app";
 
 export const ENDPOINT = `${URL}/api`;
